+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 #### Bug Fixes
 - [Predicates / Gates using closures to parameters will cause unexpected behavior.](https://github.com/SAP/chevrotain/issues/221)
 
+#### Documentation
+- [Example using parametrized sub-rules.](https://github.com/SAP/chevrotain/issues/218)
+
 
 
 ## 0.11.4 (7-16-2016)